up to 9 working tests now testing various shoppingCartService methods. Also edited some of the service methods to send back unsuccessful HttpStatus codes in isntances where shoppingCarts being searched for are not found in the shoppingCartRepo

+++ src/main/java/com/liftoff/ecommerce/Service/ShoppingCartService.java
@@ -26,21 +26,29 @@ public class ShoppingCartService {
     @Autowired
     private MovieRepository movieRepository;
 
-    public ResponseEntity<?> returnCartInstances(Long customerId){
-        List<ShoppingCart> customerCart = shoppingCartRepository.findByCustomerId(customerId);
-        if(customerCart.size()>0){
-            return new ResponseEntity<>(customerCart, HttpStatus.OK);
-        } else{
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    public ResponseEntity<?> returnAllCarts() {
+        List<ShoppingCart> shoppingCarts = (List<ShoppingCart>) shoppingCartRepository.findAll();
+        if (shoppingCarts.size() > 0) {
+            return new ResponseEntity<>(shoppingCarts, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("No carts associated with that user were found", HttpStatus.NOT_FOUND);
         }
     }
 
-    public ResponseEntity<?> createNewShoppingCart(Customer customer, ShoppingCart shoppingCart){
-        ShoppingCart newCart = new ShoppingCart(shoppingCart.getMovieId(), shoppingCart.getQuantity());
-        newCart.setCustomer(customer);
-        setTotalPrice(newCart);
-        shoppingCartRepository.save(newCart);
-        return ResponseEntity.ok(HttpStatus.CREATED);
+    public ResponseEntity<?> returnCartsByCustomerId(Long customerId) {
+        List<ShoppingCart> customerCart = shoppingCartRepository.findByCustomerId(customerId);
+        if (customerCart.size() > 0) {
+            return new ResponseEntity<>(customerCart, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("No carts associated with that user were found", HttpStatus.NOT_FOUND);
+        }
+    }
+
+    public ResponseEntity<?> createNewShoppingCart(Customer customer, ShoppingCart shoppingCart) {
+        shoppingCart.setCustomer(customer);
+        setTotalPrice(shoppingCart);
+        shoppingCartRepository.save(shoppingCart);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     public ResponseEntity<?> updateQuantityInCart(Customer customer, ShoppingCart shoppingCart) {
@@ -48,38 +56,46 @@ public class ShoppingCartService {
         Long movieIdToUpdate = shoppingCart.getMovieId();
         Long newQuantity = shoppingCart.getQuantity();
         List<ShoppingCart> allCustomersCarts = shoppingCartRepository.findByCustomerId(customerId);
-        for (ShoppingCart currentCart : allCustomersCarts) {
-            if (currentCart.getMovieId().equals(movieIdToUpdate)) {
-                currentCart.setQuantity(newQuantity);
-                setTotalPrice(currentCart);
-                shoppingCartRepository.save(currentCart);
-                break;
+        if(allCustomersCarts.size()>0) {
+            for (ShoppingCart currentCart : allCustomersCarts) {
+                if (currentCart.getMovieId().equals(movieIdToUpdate)) {
+                    currentCart.setQuantity(newQuantity);
+                    setTotalPrice(currentCart);
+                    shoppingCartRepository.save(currentCart);
+                    break;
+                }
             }
+            return new ResponseEntity<>(HttpStatus.ACCEPTED);
+        } else{
+            return new ResponseEntity<>("No carts associated with that user were found",HttpStatus.NOT_FOUND);
         }
-        return ResponseEntity.ok(HttpStatus.ACCEPTED);
     }
 
-    public ResponseEntity<?> removeItemFromCart(Customer customer, ShoppingCart shoppingCart){
+    public ResponseEntity<?> removeItemFromCart(Customer customer, ShoppingCart shoppingCart) {
         Long customerId = customer.getId();
         Long movieIdToRemove = shoppingCart.getMovieId();
         List<ShoppingCart> allCustomersCarts = shoppingCartRepository.findByCustomerId(customerId);
-        for(ShoppingCart currentCart:allCustomersCarts){
-            if(currentCart.getMovieId().equals(movieIdToRemove)){
-                Long cartId = currentCart.getCartId();
-                shoppingCartRepository.deleteById(cartId);
-                break;
+        if(allCustomersCarts.size()>0) {
+            for (ShoppingCart currentCart : allCustomersCarts) {
+                if (currentCart.getMovieId().equals(movieIdToRemove)) {
+                    Long cartId = currentCart.getCartId();
+                    shoppingCartRepository.deleteById(cartId);
+                    break;
+                }
             }
+            return new ResponseEntity<>(HttpStatus.ACCEPTED);
+        } else{
+            return new ResponseEntity<>("No carts associated with that user were found",HttpStatus.NOT_FOUND);
         }
-        return ResponseEntity.ok(HttpStatus.ACCEPTED);
     }
 
-    public ResponseEntity<?> removeAllItemsFromCart(Customer customer){
+    public ResponseEntity<?> removeAllItemsFromCart(Customer customer) {
         List<ShoppingCart> allCustomersCarts = shoppingCartRepository.findByCustomerId(customer.getId());
         shoppingCartRepository.deleteAll(allCustomersCarts);
-        return ResponseEntity.ok(HttpStatus.ACCEPTED);
+        return new ResponseEntity<>(HttpStatus.ACCEPTED);
     }
 
-    public void setTotalPrice(ShoppingCart shoppingCart){
+    public void setTotalPrice(ShoppingCart shoppingCart) {
         Long movieId = shoppingCart.getMovieId();
         Long quantity = shoppingCart.getQuantity();
         Optional<Movie> movie = movieRepository.findById(movieId);
@@ -88,10 +104,5 @@ public class ShoppingCartService {
 
         DecimalFormat decimalFormat = new DecimalFormat("0.00");
         shoppingCart.setTotalPrice(Double.parseDouble(decimalFormat.format(totalPrice)));
-    }
-
-    public ResponseEntity<?> returnAllCarts(){
-        List<ShoppingCart> shoppingCarts = (List<ShoppingCart>) shoppingCartRepository.findAll();
-        return new ResponseEntity<>(shoppingCarts, HttpStatus.OK);
     }
 }
